+++ src/constants/index.ts
@@ -3,3 +3,4 @@ export { FileExtensionConst } from './file-extension/file-extension';
 export { GlobalConst } from './global/global';
 export { CommandConst } from './command/command';
 export { RTContextConst } from './vscode-rt-context/vscode-rt-context';
+export { ExplorerConst } from './explorer/explorer';