+++ src/explorer/system-task/system-task.ts
@@ -1,19 +1,23 @@
 import { ExtensionContext, window } from 'vscode';
-import { ExplorerConst } from '../../constants/explorer/explorer';
+import { ExplorerConst } from '../../constants';
+import { IPSData, ModelingExplorer } from '../../interface';
 import { SystemTaskProvider } from './provider/system-task';
 
 /**
  * 系统运行任务
  *
  * @author chitanda
- * @date 2021-12-13 18:12:27
+ * @date 2021-12-22 15:12:53
  * @export
  * @class SystemTask
+ * @implements {ModelingExplorer<IPSData>}
  */
-export class SystemTask {
-  constructor(context: ExtensionContext) {
-    const treeDataProvider = new SystemTaskProvider(context);
-    const treeView = window.createTreeView(ExplorerConst.SYSTEM_TASK, { treeDataProvider, showCollapseAll: true });
-    context.subscriptions.push(treeView);
+export class SystemTask implements ModelingExplorer<IPSData> {
+  readonly treeDataProvider = new SystemTaskProvider(this.context);
+
+  readonly treeView = window.createTreeView(ExplorerConst.SYSTEM_TASK, { treeDataProvider: this.treeDataProvider, showCollapseAll: true });
+
+  constructor(protected readonly context: ExtensionContext) {
+    context.subscriptions.push(this.treeView);
   }
 }