+++ src/interface/index.ts
@@ -1,5 +1,6 @@
 export { GlobalContextData } from './global-context-data/global-context-data';
 export { WSMessage } from './ws-message/ws-message';
 export { CliAPIData } from './cli-api-data/cli-api-data';
+export { ModelingExplorer } from './modeling-explorer/modeling-explorer';
 // 默认标准请求回来的数据
 export type IPSData = Record<string, any>;